+++ src/edu/stuy/speed/JoeSpeed.java
@@ -11,6 +11,6 @@ package edu.stuy.speed;
 public interface JoeSpeed {
     final double ENCODER_CODES_PER_REV = 360.0;
     final double ENCODER_RPM_PER_PULSE = 60 / ENCODER_CODES_PER_REV;
-    int getRPM ();
-    void setRPM (int rpm);
+    double getRPM ();
+    void setRPM (double rpm);
 }
